requirements.txt
aiohttp>=3.9.0
websockets>=12.0
py-clob-client>=0.15.0
python-dotenv>=1.0.0
numpy>=1.24.0
colorlog>=6.7.0
eth-account>=0.10.0
web3>=6.0.0
certifi>=2023.0.0
uvloop>=0.19.0;sys_platform!='win32'
